+++ cfblog/__init__.py
@@ -1,5 +1,4 @@
-__author__ = 'vinay'
-
+# coding=utf-8
 from cfblog.middleware import Middleware
 
 __all__ = ["Middleware"]