+++ cfblog/signals.py
@@ -1,14 +1,15 @@
+# coding=utf-8
 from django.dispatch import Signal
 
 """
-    Any handler of pre_publish_signal will return two values.
-    And the results of all of these handlers will be available as a list.
-    Frist element of the tuple should be boolean or None and
-    the second one should be string.
-    True for Success,
-    False for Errors,
+    pre_publish_signal handlers should return two values.
+    Frist value should be boolean or None
+    True for Success
+    False for Errors
     None for Warnings
-    e.g. [(True, "Success",), (False, "Tag Name: Error Message...",)...]
+
+    Second value should be the corresponding message.
+    e.g. [(True, "Success",), (False, "Invalid html",)...]
 """
 pre_publish_signal = Signal(providing_args=["cms_page"])
 post_publish_signal = Signal(providing_args=["cms_page"])