+++ cfblog/tests.py
@@ -1,71 +1,124 @@
-import json
+# coding=utf-8
 import datetime
+import json
 
-from django.test import TestCase, Client
-from django.apps import apps
-from django.conf import settings
+from django.contrib.auth import get_user_model
+from django.contrib.auth.models import Permission
+from django.test import SimpleTestCase, TestCase
 
 from .models import Content
-from .utils import parse_cms_template, NAMESPACE_DELIMITER
-
-__author__ = 'vinay'
+from .utils import NAMESPACE_DELIMITER, parse_cms_template
 
 
-class CMSTests(TestCase):
-    def test_basic_features(self):
-        html = """
+class BaseTests(SimpleTestCase):
+    cms_context = {}
+
+    def setUp(self):
+        self.html = """
             <html>
                 <head>
                 </head>
                 <body>
-                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">hello</h1>
+                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
+                        hello
+                    </h1>
                 </body>
             </html>
         """
-        out = parse_cms_template(html, {})
-        self.assertHTMLEqual(html, out)
 
-        out = parse_cms_template(html, {"h1": "foo"})
+    @property
+    def output(self):
+        return parse_cms_template(self.html, self.cms_context, public=True)
+
+
+class BasicTagTests(BaseTests):
+
+    def test_empty_context(self):
+        self.cms_context = {}
+        self.assertHTMLEqual(self.html, self.output)
+
+    def test_data_cms_content(self):
+        self.cms_context = {"h1": "foo"}
         self.assertHTMLEqual(
-            out, """
-                <html>
-                    <head>
-                    </head>
-                    <body>
-                        <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">foo</h1>
-                    </body>
-                </html>
             """
-        )
-
-        self.assertHTMLEqual(
-            parse_cms_template(html, {"id": "new_id"}),
-            """
-                <html>
-                    <head>
-                    </head>
-                    <body>
-                        <h1 data-cms-content="h1" data-cms-attr="id:id" id="new_id">hello</h1>
-                    </body>
-                </html>
-            """
-        )
-
-    def test_advanced_features(self):
-        html = """
             <html>
                 <head>
                 </head>
                 <body>
-                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">hello</h1>
+                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
+                        foo
+                    </h1>
                 </body>
             </html>
-        """
-        out = parse_cms_template(
-            html,
-            {'h1': '<div data-cms-content="text"></div>'}
+            """,
+            self.output
         )
-        self.assertHTMLEqual(out, """
+
+    def test_data_cms_attr(self):
+        self.cms_context = {"id": "new_id"}
+        self.assertHTMLEqual(
+            """
+            <html>
+                <head>
+                </head>
+                <body>
+                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="new_id">
+                        hello
+                    </h1>
+                </body>
+            </html>
+            """,
+            self.output
+        )
+
+    def test_data_cms_replace(self):
+        self.html = """
+            <html>
+                <head>
+                </head>
+                <body>
+                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig"
+                        data-cms-replace>
+                    <div data-cms-content="content" data-cms-attr="id:div_id"
+                        id="replaced">
+                            new content
+                    </div>
+                    </h1>
+                </body>
+            </html>
+            """
+        self.cms_context = {}
+        self.assertHTMLEqual(
+            """
+            <html>
+                <head>
+                </head>
+                <body>
+                </body>
+            </html>
+            """,
+            self.output
+        )
+        self.cms_context = {'h1': '<div>replaced h1</div>'}
+        self.assertHTMLEqual(
+            """
+            <html>
+                <head>
+                </head>
+                <body>
+                    <div>replaced h1</div>
+                </body>
+            </html>
+            """,
+            self.output
+        )
+
+
+class MultiLevelTagTests(BaseTests):
+    def test_single_level_data(self):
+        self.cms_context = {'h1': '<div data-cms-content="text"></div>'}
+        self.assertHTMLEqual(
+            """
             <html>
                 <head>
                 </head>
@@ -75,14 +128,17 @@ class CMSTests(TestCase):
                     </h1>
                 </body>
             </html>
-        """)
-
-        out = parse_cms_template(
-            html,
-            {'h1': '<div data-cms-content="text"></div>',
-             'h1' + NAMESPACE_DELIMITER + 'text': 'replaced'}
+            """,
+            self.output
         )
-        self.assertHTMLEqual(out, """
+
+    def test_multi_level_data(self):
+        self.cms_context = {
+            'h1': '<div data-cms-content="text"></div>',
+            'h1' + NAMESPACE_DELIMITER + 'text': 'replaced'
+        }
+        self.assertHTMLEqual(
+            """
             <html>
                 <head>
                 </head>
@@ -92,14 +148,17 @@ class CMSTests(TestCase):
                     </h1>
                 </body>
             </html>
-        """)
-
-        out = parse_cms_template(
-            html,
-            {'h1': '<div data-cms-attr="id:div_id" id="default">hi</div>',
-             'h1' + NAMESPACE_DELIMITER + 'div_id': 'replaced'}
+            """,
+            self.output
         )
-        self.assertHTMLEqual(out, """
+
+    def test_single_level_complete_data(self):
+        self.cms_context = {
+            'h1': '<div data-cms-attr="id:div_id" id="default">hi</div>',
+            'h1' + NAMESPACE_DELIMITER + 'div_id': 'replaced'
+        }
+        self.assertHTMLEqual(
+            """
             <html>
                 <head>
                 </head>
@@ -109,181 +168,255 @@ class CMSTests(TestCase):
                     </h1>
                 </body>
             </html>
-        """)
-
-        out = parse_cms_template(
-            html,
-            {'h1': '<div data-cms-content="content" data-cms-attr="id:div_id" id="default">hi</div>',
-             'h1' + NAMESPACE_DELIMITER + 'div_id': 'replaced',
-             'h1' + NAMESPACE_DELIMITER + 'content': 'new content'}
+            """,
+            self.output
         )
-        self.assertHTMLEqual(out, """
+
+    def test_multi_level_complete_data(self):
+        self.cms_context = {
+            'h1': (
+                '<div data-cms-content="content" '
+                'data-cms-attr="id:div_id" id="default">hi</div>'
+            ),
+            'h1' + NAMESPACE_DELIMITER + 'div_id': 'replaced',
+            'h1' + NAMESPACE_DELIMITER + 'content': 'new content'
+        }
+        self.assertHTMLEqual(
+            """
             <html>
                 <head>
                 </head>
                 <body>
                     <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
-                    <div data-cms-content="content" data-cms-attr="id:div_id" id="replaced">new content</div>
+                    <div data-cms-content="content" data-cms-attr="id:div_id"
+                        id="replaced">
+                            new content
+                    </div>
                     </h1>
                 </body>
             </html>
-        """)
+            """,
+            self.output
+        )
 
+
+class CMSTests(BaseTests):
     def test_snippet_insertions(self):
-        html = """
+        self.cms_context = {
+            'h1': '<div data-cms-include="namespace:cms_templates/snippet">'
+                  '</div>'
+        }
+        self.assertHTMLEqual(
+            """
             <html>
                 <head>
                 </head>
                 <body>
-                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">hello</h1>
+                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
+                    <div data-cms-include="cms_templates/snippet"
+                        data-cms-namespace="namespace">
+                            hehehe
+                            <div data-cms-content="md:text">Hi</div>
+                    </div>
+                    </h1>
                 </body>
             </html>
+            """,
+            self.output
+        )
+
+    def test_snippet_insertions_data(self):
+        self.cms_context = {
+            'h1': '<div data-cms-include="namespace:cms_templates/snippet">'
+                  '</div>',
+            NAMESPACE_DELIMITER.join(
+                ('h1', 'namespace', 'text')
+            ): '##Heading##'
+        }
+        self.assertHTMLEqual(
+            """
+            <html>
+                <head>
+                </head>
+                <body>
+                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
+                    <div data-cms-include="cms_templates/snippet"
+                        data-cms-namespace="namespace">
+                            hehehe
+                            <div data-cms-content="md:text">
+                                <h2>Heading</h2>
+                            </div>
+                    </div>
+                    </h1>
+                </body>
+            </html>
+            """,
+            self.output
+        )
+
+        self.cms_context = {
+            'h1': '<div '
+                  'data-cms-include="namespace:cms_templates/snippet_2">'
+                  '</div>',
+            NAMESPACE_DELIMITER.join(
+                ('h1', 'namespace', 'text2')
+            ): '##Heading##'
+        }
+        self.assertHTMLEqual(
+            """
+            <html>
+                <head>
+                </head>
+                <body>
+                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
+                    <div data-cms-include="cms_templates/snippet_2"
+                        data-cms-namespace="namespace">
+                        <div data-cms-content="md:text2"><h2>Heading</h2></div>
+                    </div>
+                    </h1>
+                </body>
+            </html>
+            """,
+            self.output
+        )
+
+    def test_markdown(self):
+        self.html = """
+        <html>
+            <head>
+            </head>
+            <body>
+                <h1 data-cms-content="md:h1" data-cms-attr="id:id" id="orig">
+                    hello
+                </h1>
+            </body>
+        </html>
         """
-        out = parse_cms_template(
-            html,
-            {'h1': '<div data-cms-include="namespace:cms_templates/snippet"></div>'}
-        )
-        self.assertHTMLEqual(out, """
+        self.cms_context = {
+            'h1': '##Heading##\n'
+                  '<div '
+                  'data-cms-include="namespace:cms_templates/snippet_2">'
+                  '</div>',
+            NAMESPACE_DELIMITER.join(
+                ('h1', 'namespace', 'text2')
+            ): '##Heading##'
+        }
+        self.assertHTMLEqual(
+            """
             <html>
                 <head>
                 </head>
                 <body>
-                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
-                    <div data-cms-include="cms_templates/snippet" data-cms-namespace="namespace">
-                    hehehe
-                    <div data-cms-content="md:text">Hi</div>
-                    </div>
-                    </h1>
-                </body>
-            </html>
-        """)
-
-        out = parse_cms_template(
-            html,
-            {'h1': '<div data-cms-include="namespace:cms_templates/snippet"></div>',
-             NAMESPACE_DELIMITER.join(('h1', 'namespace', 'text')): '##Heading##'}
-        )
-        self.assertHTMLEqual(out, """
-            <html>
-                <head>
-                </head>
-                <body>
-                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
-                    <div data-cms-include="cms_templates/snippet" data-cms-namespace="namespace">
-                    hehehe
-                    <div data-cms-content="md:text"><h2>Heading</h2></div>
-                    </div>
-                    </h1>
-                </body>
-            </html>
-        """)
-
-        out = parse_cms_template(
-            html,
-            {'h1': '<div data-cms-include="namespace:cms_templates/snippet_2"></div>',
-             NAMESPACE_DELIMITER.join(('h1', 'namespace', 'text2')): '##Heading##'}
-        )
-        self.assertHTMLEqual(out, """
-            <html>
-                <head>
-                </head>
-                <body>
-                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
-                    <div data-cms-include="cms_templates/snippet_2" data-cms-namespace="namespace">
-                    <div data-cms-content="md:text2"><h2>Heading</h2></div>
-                    </div>
-                    </h1>
-                </body>
-            </html>
-        """)
-
-        out = parse_cms_template(
-            html,
-            {'h1': '##Heading##\n'
-                   '<div data-cms-include="namespace:cms_templates/snippet_2"></div>',
-             NAMESPACE_DELIMITER.join(('h1', 'namespace', 'text2')): '##Heading##'}
-        )
-        self.assertHTMLEqual(out, """
-            <html>
-                <head>
-                </head>
-                <body>
-                    <h1 data-cms-content="h1" data-cms-attr="id:id" id="orig">
-                    ##Heading##
-                    <div data-cms-include="cms_templates/snippet_2" data-cms-namespace="namespace">
-                    <div data-cms-content="md:text2"><h2>Heading</h2></div>
-                    </div>
-                    </h1>
-                </body>
-            </html>
-        """)
-
-        html = """
-            <html>
-                <head>
-                </head>
-                <body>
-                    <h1 data-cms-content="md:h1" data-cms-attr="id:id" id="orig">hello</h1>
-                </body>
-            </html>
-        """
-        out = parse_cms_template(
-            html,
-            {'h1': '##Heading##\n'
-                   '<div data-cms-include="namespace:cms_templates/snippet_2"></div>',
-             NAMESPACE_DELIMITER.join(('h1', 'namespace', 'text2')): '##Heading##'}
-        )
-        self.assertHTMLEqual(out, """
-            <html>
-                <head>
-                </head>
-                <body>
-                    <h1 data-cms-content="md:h1" data-cms-attr="id:id" id="orig">
+                    <h1 data-cms-content="md:h1" data-cms-attr="id:id"
+                        id="orig">
                     <h2>Heading</h2>
-                    <div data-cms-include="cms_templates/snippet_2" data-cms-namespace="namespace">
-                    <div data-cms-content="md:text2"><h2>Heading</h2></div>
+                    <div data-cms-include="cms_templates/snippet_2"
+                        data-cms-namespace="namespace">
+                            <div data-cms-content="md:text2">
+                                <h2>Heading</h2>
+                            </div>
                     </div>
                     </h1>
                 </body>
             </html>
-        """)
-
-    def test_versioning(self):
-        client = Client()
-        app_label, model = settings.AUTH_USER_MODEL.split('.', 1)
-        User = apps.get_model(app_label, model)
-        user = User.objects.create_superuser(**{
-            User.USERNAME_FIELD: 'test-golbfc',
-            'email': 'test@test.com',
-            'password': 'test123'}
+            """,
+            self.output
         )
+
+
+class RegressionTests(TestCase):
+
+    @classmethod
+    def setUpTestData(cls):
+        User = get_user_model()
+        editor_credentials = {
+            User.USERNAME_FIELD: 'editor',
+            'password': 'test123'
+        }
+        author_credentials = {
+            User.USERNAME_FIELD: 'author',
+            'password': 'test123'
+        }
+        editor = User.objects.create_user(is_active=True, **editor_credentials)
+        author = User.objects.create_user(is_active=True, **author_credentials)
+        edit_permission = Permission.objects.get(
+            codename='change_content',
+            content_type__app_label='cfblog'
+        )
+        publish_permission = Permission.objects.get(
+            codename='can_publish',
+            content_type__app_label='cfblog'
+        )
+        author.user_permissions.add(publish_permission)
+        editor.user_permissions.add(edit_permission)
+        cls.author = author
+        cls.editor = editor
+        cls.author_client = cls.client_class()
+        cls.author_client.login(**author_credentials)
+        cls.editor_client = cls.client_class()
+        cls.editor_client.login(**editor_credentials)
+
+    def setUp(self):
         cms_page, _ = Content.objects.get_or_create(
             url='/test-golbfc/',
             template="cms_templates/template_1.html",
             category_id=1,
-            author=user
+            author=self.author
         )
+        self.cms_page = cms_page
 
-        content = """
-                    {
-                      "body":"ABCDEFGHIJKLMNOPQRSTUVWXYZ",
-                      "body_attr":"random data",
-                      "decrp":"Hi!!",
-                      "title":"Hello World!!!"
-                    }
-                  """
-        t = (cms_page.modified_on + datetime.timedelta(minutes=5)).isoformat()
-        data = {
-                'auth_data': content,
-                'draft_modified': t,
-                'cms_page_id': cms_page.id
-                }
-        self.assertTrue(
-            client.login(**{
-                User.USERNAME_FIELD: 'test-golbfc', 'password': 'test123'
-            })
+    def test_unpublished_url(self):
+        self.cms_page.status = Content.DRAFT
+        self.cms_page.save()
+        response = self.client.get(self.cms_page.url)
+        self.assertEqual(response.status_code, 404)
+        response = self.editor_client.get(self.cms_page.url)
+        self.assertEqual(response.status_code, 200)
+        response = self.author_client.get(self.cms_page.url)
+        self.assertEqual(response.status_code, 200)
+
+    def test_published_url(self):
+        self.cms_page.status = Content.PUBLIC
+        self.cms_page.save()
+        response = self.client.get(self.cms_page.url)
+        self.assertEqual(response.status_code, 200)
+        response = self.editor_client.get(self.cms_page.url)
+        self.assertEqual(response.status_code, 200)
+        response = self.author_client.get(self.cms_page.url)
+        self.assertEqual(response.status_code, 200)
+
+    def test_editor_cannot_publish(self):
+        response1 = self.editor_client.post(
+            path='/cms/ajax/save/publish/',
+            data={},
+            HTTP_X_REQUESTED_WITH='XMLHttpRequest'
         )
-        response1 = client.post(
+        self.assertEqual(response1.status_code, 403)
+
+    def test_author_can_publish(self):
+        response1 = self.author_client.post(
+            path='/cms/ajax/save/publish/',
+            data={},
+            HTTP_X_REQUESTED_WITH='XMLHttpRequest'
+        )
+        self.assertNotEqual(response1.status_code, 403)
+
+    def test_versioning(self):
+
+        content = json.dumps({
+            "body": "ABCDEFGHIJKLMNOPQRSTUVWXYZ",
+            "body_attr": "random data",
+            "decrp": "Hi!!",
+            "title": "Hello World!!!"
+        })
+        draft_time = self.cms_page.modified_on + datetime.timedelta(minutes=5)
+        data = {
+            'auth_data': content,
+            'draft_modified': draft_time.isoformat(),
+            'cms_page_id': self.cms_page.id
+        }
+
+        response1 = self.author_client.post(
             path='/cms/ajax/save/publish/',
             data=data,
             HTTP_X_REQUESTED_WITH='XMLHttpRequest'
@@ -291,8 +424,9 @@ class CMSTests(TestCase):
         self.assertEqual(response1.status_code, 200)
         self.assertTrue(json.loads(response1.content)['success'])
 
-        data['draft_modified'] = '2016-04-07T13:45:35.322Z'
-        response2 = client.post(
+        draft_time = self.cms_page.modified_on - datetime.timedelta(minutes=5)
+        data['draft_modified'] = draft_time.isoformat()
+        response2 = self.author_client.post(
             path='/cms/ajax/save/publish/',
             data=data,
             HTTP_X_REQUESTED_WITH='XMLHttpRequest'