+++ cfblog/validators.py
@@ -1,13 +1,15 @@
+# coding=utf-8
 from __future__ import unicode_literals
-__author__ = 'vinay'
+
 import re
 
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.core.validators import RegexValidator
-from django.template.loader import get_template, select_template, TemplateDoesNotExist
-from django.utils.translation import ugettext_lazy as _
+from django.template.loader import (TemplateDoesNotExist, get_template,
+                                    select_template)
 from django.utils.lru_cache import lru_cache
+from django.utils.translation import ugettext_lazy as _
 
 
 @lru_cache()